Make the images centered vertically with the text
To move the images independently from the text, align-self is needed.
By default, the cross axis runs vertically with the main direction is
row. But that will just move the items individually up and down, not
lef and right to align it vertically with other items. The solution is
to switch the main direction so that the cross axis runs horizonatally
and then the items can be moved left and right to be aligned vertically
with other items.