Remove self-align of the center item in the cross axis
The wanted result is to align everything inside the center item
vertically. Since the main body runs from top to bottom, the center item
streches out horizontally across the cross axis. That means the center
container is as wide as the screen.

And because of that, just align-items alone can put the everything inside
the center container centered on the screen.